cleaned up the feature branch before creating a pull request. to see Cython/C exploratory work, go back to the previous commit in this branch
