+++ setup.py
@@ -1,6 +1,4 @@
-from setuptools import setup, find_packages, Extension
-from Cython.Build import cythonize
-import numpy as np
+from setuptools import setup, find_packages
 
 setup(
     name='TextEmbeddingFE',
@@ -24,12 +22,4 @@ setup(
     description='Feature extraction from text fields using text-embedding large language models',
     license='MIT',
     keywords=['text embedding', 'large language models', 'feature engineering', 'cross-validation'],
-    ext_modules=cythonize([
-        Extension(
-            name="TextEmbeddingFE.skmeans_lloyd_update",
-            sources=["TextEmbeddingFE/skmeans_lloyd_update.pyx", "TextEmbeddingFE/skmeans_lloyd_update_cpp.cpp"],
-            include_dirs=[np.get_include()]
-            ,language="c++"
-        )
-    ])
 )
